user-definable timeout reporter
Allow the user to define a lambda which will
be used to report when a timeout occurs. By default
the rack env and the exception will be inspected
and sent to STDOUT (this could be done much better).

Here's an example of using a custom handler to report
occurrences of timeouts to exceptional.io:

Rack::Timeout.reporter = lambda{|exception, env|
  ::Exceptional::Catcher.handle_with_rack(exception, env, Rack::Request.new(env))
}