don't let other middleware handle timeout exception
because of how Timeout#timeout is implemented, any code that handles
exceptions in the block being called will handle the
Rack::Timeout::AppTimeout exception that Timeout raises when the code times out.
This makes it impossible to rescue from the timeout scenario being tested
in the first place.

Wrapping the app invocation inside a thread results in Timeout
raising its exception inside the wrapper thread instead of inside the app.